# backend/internal/handler/api.gen.go
// Package handler provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package handler

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Item defines model for Item.
type Item struct {
	// ShopSpecificName Shop-specific name of the item
	ShopSpecificName string `json:"shop_specific_name"`

	// UniversalName Universal name of the item (e.g., combi-hadagi)
	UniversalName string `json:"universal_name"`
}

// RecommendationResponse defines model for RecommendationResponse.
type RecommendationResponse struct {
	// AgeInMonths Calculated age in months
	AgeInMonths int `json:"age_in_months"`

	// Items List of recommended items
	Items []Item `json:"items"`
}

// GetRecommendationParams defines parameters for GetRecommendation.
type GetRecommendationParams struct {
	// BirthDate Baby's birth date (YYYY-MM-DD)
	BirthDate openapi_types.Date `form:"birth_date" json:"birth_date"`

	// TargetShop Target shop name for specific terminology (e.g., uniqlo, nishimatsuya, akachan_honpo)
	TargetShop *string `form:"target_shop,omitempty" json:"target_shop,omitempty"`

	// TargetDate Target date to wear the baby wear (YYYY-MM-DD), defaults to today
	TargetDate *openapi_types.Date `form:"target_date,omitempty" json:"target_date,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get recommended baby wear
	// (GET /recommend)
	GetRecommendation(c *gin.Context, params GetRecommendationParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetRecommendation operation middleware
func (siw *ServerInterfaceWrapper) GetRecommendation(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRecommendationParams

	// ------------- Required query parameter "birth_date" -------------

	if paramValue := c.Query("birth_date"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument birth_date is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "birth_date", c.Request.URL.Query(), &params.BirthDate)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter birth_date: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "target_shop" -------------

	err = runtime.BindQueryParameter("form", true, false, "target_shop", c.Request.URL.Query(), &params.TargetShop)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter target_shop: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "target_date" -------------

	err = runtime.BindQueryParameter("form", true, false, "target_date", c.Request.URL.Query(), &params.TargetDate)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter target_date: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetRecommendation(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/recommend", wrapper.GetRecommendation)
}